backend/requirements.txt
fastapi 
uvicorn[standard] 
algokit-utils 
python-dotenv 
pydantic
google-generativeai